requirements.txt
PySide6>=6.6.0
supabase>=2.0.0
keyring>=24.0.0
python-dotenv>=1.0.0

